% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helpers.R
\name{calcSigma}
\alias{calcSigma}
\title{Simulation Helper: GP-kernel}
\usage{
calcSigma(X1, X2, l = 1)
}
\arguments{
\item{X1}{A vector of input}

\item{X2}{A vector of input}

\item{l}{A number of kernel window size}
}
\description{
Calculates the Gaussian Process kernel for simulation
}
