% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfm.R
\name{hfm.LF}
\alias{hfm.LF}
\title{Multivariate Hierarchical Functional Models: Latent Factors Implementation (SS and NS prior)}
\usage{
hfm.LF(y, t, X, nsegs, nLF = NULL, knots = NULL, mcmc = NULL)
}
\arguments{
\item{y}{A matrix with vectorized region measurements in columns.}

\item{t}{A vector with vectorized observed time points or segments.}

\item{X}{A regression design matrix (subject by predictor).}

\item{nsegs}{An integer vector with the number of observed segments per subject.}

\item{nLF}{Number of latent factors. Default is (nkonts * nregions)/2, specify one argument for NS prior, two arguenets for SS prior.}

\item{knots}{Number of spline knots or a vector of spline knots.}

\item{mcmc}{A list of mcmc specifications. Default is (burnin = 1000, nsim = 2000, thin = 1)}
}
\value{
A list of with model fit summaries list(fit, coef, y, t, Bs).
\itemize{
\item{fit:}{ Posterior mean fit to y in a 3d array [subject, region, segment]}
\item{coef:}{ Array containing mean posterior regression coefficients [region, spline, predictor]}
\item{gfit:}{ Posterior Fixed effects fit in a 3d array [subject, region, segment]}
\item{ifit:}{ Posterior Latent Factor model fit in a 3d array [subject, region, segment]}
\item{y:}{ Reashaped data array [subject, region, segment]}
\item{t:}{ Matrix of observed segments [subject, segments]}
\item{Bs:}{ Matrix of spline polynomials used in the model}
\item{X:}{ Covariate matrix}
\item{nsim:}{ Number of MCMC samples}
}
}
\description{
This function is a wrapper for a Cpp Gibbs Sampler. Multivariate functional observations, for examples t time points observed over p regions simultaneously, are projected onto B-spline bases, and further modeled by means of a latent factors encoded hierarchical priors. We implement both a Strongly Separable and Non-Separable priors.
}
\details{
For region r, subject i, let y_i(t) ...
}
\examples{


}
