% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfm.R
\name{hfm.NB}
\alias{hfm.NB}
\title{Multivariate Hierarchical Functional Models: Matrix Normal Implementation (NB prior)}
\usage{
hfm.NB(y, t, X, nsegs, knots = NULL, mcmc = NULL, spatial = FALSE)
}
\arguments{
\item{y}{A matrix with vectorized region measurements in columns.}

\item{t}{A vector with vectorized observed time points or segments.}

\item{X}{A regression design matrix (subject by predictor).}

\item{nsegs}{An integer vector with the number of observed segments per subject.}

\item{knots}{Number of spline knots or a vector of spline knots}

\item{mcmc}{A list of mcmc specifications. Default is (burnin = 1000, nsim = 2000, thin = 1)}
}
\value{
A list of with model fit summaries list(fit, coef, y, t, Bs).
\itemize{
\item{fit:}{ Posterior mean fit to y in a 3d array [subject, region, segmenr]}
\item{coef:}{ Array containing mean posterior regression coefficients [time, predictor, region]}
\item{coef2:}{ Array containing second moment of posterior regression coefficients [time, predictor, region]}
\item{y:}{ Reashaped data array [subject, region, segment]}
\item{t:}{ Matrix of observed segments [subject, segments]}
\item{Bs:}{ Matrix of spline polynomials used in the model}
\item{X:}{ Covariate matrix}
\item{nsim:}{ Number of MCMC samples}
}
}
\description{
This function is a wrapper for a Cpp Gibbs Sampler. Multivariate functional observations, for examples t time points observed over p regions simultaneously, are projected onto B-spline bases, and further modeled by means of a matrix normal prior.
}
\details{
For region r, subject i, let y_i(t) ...
}
\examples{


}
