% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.HFM.R
\name{sim.HFM}
\alias{sim.HFM}
\title{Model Generated Simulations}
\usage{
sim.HFM(nsub = 10, nreg = 6, tt = NULL, type = "sep", pr.miss = 0.2,
  SNR = 1, nLF = 5, nbase = 10, sp_rho = runif(1) * 0.9)
}
\description{
This simulation function generates functional observations from the HFMM models (separable and non-separable).
}
