% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{bands}
\alias{bands}
\title{Simultaneous posterior bands from hierarchical regionally referenced varying coefficient model}
\usage{
bands(fit, beta, alpha = NULL)
}
\arguments{
\item{fit}{An object returned from \code{hfm}}

\item{beta}{A matrix of mcmc samples from \code{hfm}}

\item{alpha}{Value between 0 and 1 for the (1-alpha)\% Credible Band.}
}
\value{
An array with simultaneous bands for the varying coefficients in fit
}
\description{
This function is a wrapper for a Cpp Gibbs Sampler. Region-specific regression is modeled through a varying coefficient functions.
}
\details{
For region r, subject i, let y_i(t) ...
}
\examples{


}
